requirements.txt
requests
openai
python-dotenv
groq